refactor: rename compat3 to string_utils
